PIF Registry refactor
