+++ handlers/submission_handler.py
@@ -26,6 +26,7 @@ from praw.models import Comment, Submission  # type: ignore[import-untyped]
 
 from handlers.comment_handler import handle_comment
 from pifs import pif_builder
+from pifs.registry import known_pif_types
 from utils.pif_storage import pif_exists, save_pif
 
 
@@ -73,7 +74,7 @@ def has_latherbot_pif_command(submission: Submission) -> bool:
         if line.strip().startswith("latherbot"):
             logging.info('Post "%s" MIGHT have a LatherBot command', submission.title)
             parts = line.split()
-            if parts[1] in pif_builder.known_pif_types():
+            if parts[1] in known_pif_types():
                 logging.info('Submission "%s" is a PIF!', submission.title)
                 return True
