+++ pifs/battleship_pif.py
@@ -15,7 +15,7 @@ from typing import Any
 from praw.models import Comment, Redditor  # type: ignore[import-untyped]
 
 from pifs.base_pif import BasePIF
-from pifs.pif_builder import register_pif
+from pifs.registry import register_pif
 
 instructionTemplate = """
 Welcome to {}'s Battleship PIF (managed by LatherBot).
