+++ pifs/geo_pif.py
@@ -16,7 +16,7 @@ from geopy.geocoders import Nominatim  # type: ignore[import-untyped]
 from praw.models import Comment, Redditor  # type: ignore[import-untyped]
 
 from pifs.base_pif import BasePIF
-from pifs.pif_builder import register_pif
+from pifs.registry import register_pif
 from utils.reddit_helper import get_comment, user_agent
 
 instructionTemplate = """
