+++ pifs/karma_only_pif.py
@@ -12,7 +12,7 @@ from typing import Any
 from praw.models import Comment, Redditor  # type: ignore[import-untyped]
 
 from pifs.base_pif import BasePIF
-from pifs.pif_builder import register_pif
+from pifs.registry import register_pif
 from utils.karma_calculator import calculate_karma, formatted_karma
 from utils.reddit_helper import get_submission
